Move auth, creds, and multipart into the utils package.
