+++ alburnum/maas/flesh/__init__.py
@@ -6,19 +6,23 @@
 
 __all__ = []
 
+import argcomplete
+import argparse
+import collections
+import colorclass
+import csv
+import enum
+import json
+import sys
+import terminaltables
+import yaml
 from abc import (
     ABCMeta,
     abstractmethod,
 )
-import argparse
-import collections
-import csv
-import enum
 from io import StringIO
-import json
 from operator import itemgetter
 from os import environ
-import sys
 from textwrap import fill
 
 from alburnum.maas import (
@@ -26,13 +30,9 @@ from alburnum.maas import (
     utils,
     viscera,
 )
-from alburnum.maas.auth import obtain_credentials
 from alburnum.maas.bones import CallError
-from alburnum.maas.creds import Credentials
-import argcomplete
-import colorclass
-import terminaltables
-import yaml
+from alburnum.maas.utils.auth import obtain_credentials
+from alburnum.maas.utils.creds import Credentials
 
 
 def check_valid_apikey(_1, _2, _3):  # TODO
